feat: make StartDate optional in PrescribedDose; fix family/supplement names display
- Make StartDate nullable (optional) in PrescribedDose model, DB, and views
- Replace Dapper multi-mapping with flat query using FamilyMemberName/SupplementName
- Fix family member and supplement names showing blank in PrescribedDose Index
- Update Create/Edit/Delete/Index views for nullable StartDate and new name properties
- Update ReportingController to handle nullable StartDate in active dose filter
- All 15 unit tests and 15 Playwright tests pass

+++ VitaTrack.Infrastructure/Data/DbInit.cs
@@ -58,7 +58,7 @@ namespace VitaTrack.Infrastructure.Data
             if (columnExists == 0)
             {
                 db.Execute(@"
-                    ALTER TABLE PrescribedDoses ADD COLUMN FrequencyPerDay INTEGER NOT NULL DEFAULT 1;
+                    ALTER TABLE PrescribedDoses ADD COLUMN FrequencyPerDay REAL NOT NULL DEFAULT 1.0;
                 ");
             }
